+++ src/components/AdminPanel/ErrorModal/ErrorModal.js
@@ -1,0 +1,16 @@
+import React from "react";
+import styles from "./ErrorModal.module.css";
+import { ReactComponent as Error } from "../../../assets/admin/auth/error.svg";
+
+const ErrorModal = ({ message }) => {
+  return (
+    <div className={styles.error}>
+      <div>
+        <Error />
+      </div>
+      <p>{message}</p>
+    </div>
+  );
+};
+
+export default ErrorModal;
