+++ src/pages/AdminPanel/Auth/Forgot/ForgotPass.jsx
@@ -1,6 +1,6 @@
 import React from "react";
 import styles from "../Login.module.css";
-import ForgotForm from "./ForgotForm";
+import ForgotForm from "../../../../components/AdminPanel/Auth/Forgot/ForgotForm";
 
 const ForgotPass = () => {
   return (
