+++ src/pages/AdminPanel/Auth/Reset/ResetPass.jsx
@@ -1,6 +1,6 @@
 import React from "react";
 import styles from "../Login.module.css";
-import ResetForm from "./ResetForm";
+import ResetForm from "../../../../components/AdminPanel/Auth/Reset/ResetForm";
 
 const ResetPass = () => {
   return (
